[API] Visitor methods now have empty default.
The methods of visitors created with our `M_DECLARE_VISITOR` were purely
virtual. This forced the user of `M_DECLARE_VISITOR` to
override/implement every method.  This commit provides an empty default
for the methods.  This allows a user of a visitor created with
`M_DECLARE_VISITOR` to override only some methods of the visitor.