[Sema] Don't perform DDL during semantic analysis.
Before this patch, the `Sema` did perform certain DDL commands as part
of its semantic analysis.  This coupling of semantic analysis and
performing DDL commands, however, may lead to undesired effects, e.g.
the creation of a table during analysis of a `CREATE TABLE`-statement
even when one of the declarations of an attribute has a semantic error,
e.g. the expression in the `CHECK()`-constraint is not of boolean type.
The `Sema` then raises an error and does not add that attribute to the
table.  The client receives a diagnostic error and expects the table to
*not* have been created.

Full example:
```sql
CREATE TABLE erroneous (
    id INT(4), -- OK
    err INT(4) CHECK(id - 42) -- ERROR: not of boolean type
);
```

During semantic analysis of `CREATE TABLE erroneous ...`, the table was
created but attribute `err` was not added.

This patch removes performing any DDL commands during semantic analysis.
The `Sema` only performs necessary checks but *does not alter* the DB
state.  Instead, the method `m::execute_statement()` is extended to
handle `CreateDatabaseStmt`, `UseDatabaseStmt`, and `CreateTableStmt`
and run the respective operations of the DDL.