[mutable] Use `Sema::analyze()` in `check` binary.
In oour `check` utility binary, call new `Sema::analyze()` instead of
old `Sema::operator()()` to perform semantic analysis.  Then, if the
returned `DatabaseCommand` is a `DDLCommand`, execute it to run DDL as
part of `check`.  (This is now required, as the `Sema` *does not*
perform DDL commands as part of its semantic analysis anymore.  See
prior commit 908b37c7 for details.)