Remove workflow rule for pushed tags
Before this change, every time a tag was pushed to GitLab, a pipeline for the tagged commit would be executed. This pipeline execution is redundant since every commit should be either part of the main branch or of a MR and therefore a pipeline was already executed for the tagged commit.