[Benchmark] Add `NOT NULL` constraints.
Our benchmarks did not make use of the `NOT NULL` constraints.  However,
these significantly reduce the amount of generated and executed code and
hence improve performance.  Therefore, add the missing `NOT NULL`
constraints to all attributes of benchmark suites `operators` and
`tpc-h`.  Note, that TPC-H explicitly *allows* for adding the `NOT NULL`
constraint to *all* attributes.