[Test] Report the CLI args of binary invocations.
Report the CLI args that are being passed to the binaries that are
invoked by the test script.

We will discard some of the first CLI args, as they are the same for all
invocations.  However, the amount of CLI args being discarded from the
front is currently hard-coded.  Perhaps we can at least set a number per
binary, e.g. 5 for `shell`.