[Wasm] Do not write constants into result set but add them for printing.
Materialized data, e.g. intermediate buffers as well as the result set,
do not have to contain constants as they either have no use (if they occur
in a subquery) or can be added when printing the result (if they occur
in the outermost SELECT).  This commit adds the support for arbitrary
constants (not only NULL as before) but still does not support constant
expressions as we are currently not able to interpret them.