[Benchmark] Refactor benchmark system.
Instead of using a different script for each experiment and DBMS to
benchmark, implement `connectors` to these DBMS's. The connector has a
method to execute an experiment with the given parameters and returns
the measured times.

In addition, the format of the YAML files of the experiments has been
refactored to contain all the information and parameters to execute them
on each connector.

`Benchmark.py` is refactored as well to read the experiment files and
execute them on each available specified connector, with possibly
multiple configurations.

Some more minor changes:
- Benchmark script now has the option to execute one (or multiple)
  specific experiments.
- The `run_id` of each experiment run is tracked and inserted into the
  database.