[Wasm] Introduce `setup_t` and `teardown_t` to avoid incorrect ordering of callbacks.
The parent teardown code must always be placed before the own one to ensure that
`wasm::Limit` jumps to the correct code when the limit is reached.  To avoid
violating this constraint, the c'tor of `teardown_t` automatically ensures this
order.  However, this behaviour can still be bypassed if necessary (see e.g.
`wasm::Limit::execute()`) but in a more explicit way than before.