[Wasm] Fix and adapt sequential data layout accesses.
Before this commit, accesses which were compiled as single pass but
were actually executed in multiple function calls (by reexecuting the
init block with an updated tuple ID, see `Buffer::consume()`) did not
reinitilize the pointers and masks correctly.  To solve this issue, we
adapted `compile_data_layout_sequential()` s.t. it now compiles the
inits dependent of the current tuple ID.  However, the former initial
compile-time tuple ID offset was removed as it would now result in
tuple gaps of the size of this offset which would be skipped at each
execution of the init block.