[Benchmark] Fix a bug in HyPer connector.
The HyPer connector used the `multiprocessing` library to run queries in
a separate thread.  This, however, resulted in copy-on-write for the
database connection.  Since Python objects are reference counted, COW
was always triggered and the connection object duplicated.  This, then,
resulted in the child thread closing the connection properly and the
main thread trying to close the connection a second time, resulting in
error.

With this patch, we get rid of `multiprocessing` entirely.  We simply
run HyPer on all coresthe same way as on a single core: in a nested
Python interpreter, created and managed with `subprocess`.  This avoids
the nasty COW bug.