[IR] Fix multi-pred. equi-joins in Interpreter.
The old implementation for handling equi-joins in simple-hash join in
the interpreter assumed the join condition be a single equi predicate,
e.g. `A.x = B.y`. However, we do allow for conjunctions of
equi predicates and consider them equi-join conditions.

This commit implements a proper decomposition of a equi-join condition
of multiple predicates.  Consequently, we don't build a join key of a
single value anymore but of as many values as there are equi predicates
in the join condition.