[Pipenv] Roll back to Python 3.10.
The package `Python-EasyGraph`, that provides the `easygraph` module, is
incompatible with Python 3.11 (at least for writing the graph as
Graphviz dot).  Therefore, roll back to Python 3.10 and consequently
also update the installed package versions.

We now also install PyTorch, as this is (optionally) required by
`Python-EasyGraph`.  If it is not installed, `easygraph` will spit out a
bunch of warnings that cannot be silenced the usual way.