[Benchmark] Fix case handling of table names.
All connectors except mu*t*able perform case-insensitive query parsing.
This means, that both `CREATE TABLE <NAME>` and `SELECT ... FROM <NAME>`
queries had the `<NAME>` internally converted to lowercase.  This works
fine if we always use the SQL interface only.  However, for HyPer, we
use the `tableauhyperapi` for creating and disposing of tables, and that
API is case sensitive.  This led to some unexpected behavior, and a
first workaround was to convert table names to lower case before passing
into the HyPer API.

The proper solution, however, is to acknowledge that SQL -- or at least
PostgreSQL compliant parsers -- are case insensitive by default.
Therefore, we should quote table names on all uses, to make sure the
parser respects the actual case.

This commit addresses this issue by two major changes:
1. Rewrite all benchmarks to properly quote the table names and use the
   correct case.
2. Adapt the database connectors to acknowledge the case and quote table
   names whenever SQL statements are being generated.