[IR] Fix critical bug in `Table::at()`.
The implementation of `Table::at()` was incorrect.
The old implementation used `name_to_attr_` to get the internal ID and
delegate this ID to another overload of `Table::at()`.  However, if the
name is not present, calling `name_to_attr_.at()` will raise an
unhandled `std::out_of_range`.

The fix here is to *not* use `name_to_attr_` directly but to delegate
the name on to the non-`const` overload of `Table::at()` directly.  This
implementation will then access `name_to_attr_` through `find()` rather
than `at()`.