[IR] Improve schema minimization.
So far, schema minimization was limited as the `Interpreter` required an
`Operator` schema to actually describe the internal physical structure
of a `Tuple` as passed through internal `Pipeline`s.

This commit aims to minimize the `Schema` of `Operator`s in the plan as
much as possible.  This may break backend compatibility.  A next commit
will adapt the backends to the properly minimized `Schema`s.