[Backend] Update backends for minimized `Schema`.
The prior commit introduced a more rigorous minimization of `Schema`s of
`Operator`s in a plan.  This commit makes the necessary changes in the
backends to handle the new `Schema`.

So far, the `Interpreter` relied on the `Schema` not being exactly
*minimal* but simply a concatenation of the `Schema`s of both sides of a
join.  This commit fixes the implementation of the simple-hash join to
correctly use the minimized schema.  Further, we drop the requirement
that `Tuple`s be simply concatenated.