[Backend] Provide recursive `dump()` in `MatchBase`.
The prior implementation of `PhysicalOptimizer::dump_plan()` performed
the traversal of the physical plan itself.  This has the downside that
we cannot use semantic information of an operator's children to provide
additional information in the dump.

This refactored implementation of `PhysicalOptimizer::dump_plan()` now
delegates to a recursive `MatchBase::print()` that contains additional
semantic information about an operator's children.  For example, the
`SimpleHashJoin` will annotate which of its children is the build input
and which is the probe input.