[Benchmark] Measure *execution time* in operators.
Change the pattern to capture mu*t*able's *execution time* rather than
end-to-end time in the `operators` benchmark suite.
This is more fair, as we also get HyPer's execution time w/o
compile time or other preceding tasks.

To properly support measuring execution time of compiled queries *and*
end-to-end time (execution time) of interpreted queries, we extend the
benchmark schema and adapt the script to allow for *per-configuration*
patterns.