[Benchmark] Add type hints to benchmark script and connectors
We annotate each class, function, and variable with a type hint. While
python does not have static types, these type hints make it easier to
understand the whole benchmarking system.

We also define type aliases for benchmark data as well as connector
results indicating a clear structure for the underlying data. This makes
it easier to understand the types and values one is dealing with, as
nested dictionaries without type hints can be confusing.

Furthermore, typeguard is used by annotating functions and classes with
`@typechecked` in order to ensure that all variables and function return
types fulfill the specified type.