[Frontend] Remove `has_next()` from `Lexer`.
The result of `has_next()` can be one of true positive, true negative,
or false negative.  Therefore, if `has_next()` returns `false`, there is
no semantically meaningful way of using the return value.

Furthermore, it is safe to always extract the next token from the input
stream independent of the result of `has_next()`.  We adapt `consume()`
to always extract the next token.