[Parser] Simplify recovery for statements, clauses
We introduce a new template function `recover()` that given a type `T`, a
token, and a follow set, performs recovery and returns an object of `T`
that was constructed on the token.  This simplifies recovery for
statements and clauses as we no longer need to use `goto` statements.
Furthermore, we properly consume a token in the beginning of derivation
to ensure that the parser does not diverge.