[IR] Fix bug in `PlanTable::dump()`.
In `PlanTable::dump()`, we compute the width of the cost column.  This
computation yields UB when one of the cells has infinite cost.  Recall,
that cells of the `PlanTable` are initialized with infinite cost and
these costs are only overwritten later, e.g. through
`PlanTable::update()`.  Dumping the `PlanTable` before all cells were
assigned an actual cost causes said UB.