[Frontend] Resolve TODO int `Expr::get_required()`
In fact, the old implementation was incorrect in that it terminated tree
traversal with the visitor entirely whereas the new implementation
really just stops recursing further down into the arguments of the
`FnApplicationExpr` aggregation.

However, it seems we never encountered a case where that mattered.  As
an example, consider the expression
```sql
AVG(b) + a
```
where *b* is vectorial and *a* is scalar.  In the old implementation,
when `AVG` was visited, the entire tree traversal was stopped and *a*
would never be visited.  This was an error.