[IR] Improve `get_aggregates()`.
We had an outstanding TODO in a stale branch, where we wanted to skip
traversing the subtree of a `FnApplicationExpr` aggregation.  Now that
we have the respective exception `visit_skip_subtree` and the logic in
the AST visitors, we can resolve this TODO and improve the performance
of `get_aggregates()`.

This change should not change the outcome of `get_aggregates()`.  As
aggregate functions take vectorial arguments and produce scalar results,
they could have never occur nested, e.g. `MIN(MAX(x))` is invalid.
Hence, not skipping subtree traversal was not an error before.