[Schema] Add `MultiVersioningTable` decorator
In this commit, the `MultiVersioningTable` decorator is introduced. Currently, it incorporates two attributes, `$ts_begin` and `$ts_end`, designed to store timestamps that denote the validity period of a tuple. However, at this stage, these attributes do not have any functionality beyond their storage.

Additionally, a new attribute constraint, `HIDDEN`, is added. This constraint serves to mark attributes as internal and, consequently, invisible to the user. To support this addition, a custom iterator for `Table` is introduced, allowing for variable skipping of hidden or visible attributes during iteration. Several new methods are introduced to maintain the same `Table` access functionality as before the introduction of hidden attributes.