[Catalog] Add the `TableFactory` Component.
This commit introduces the `TableFactory` as a component to the `Catalog`. It departs from the conventional practice of registering `TableFactory` instances in a `ComponentSet`. Instead, it establishes a single `TableFactory` object named `table_factory_`, which functions as the implicit default `TableFactory`. This default `table_factory_` is by default an undecorated `ConcreteTableFactory`.

Additionally, the `Catalog` now maintains a `ComponentSet` called `table_factory_decorators_`, which contains callback functions, each corresponding to a concrete implementation of `TableFactoryDecorator`. These callback functions are designed for decorating a given `TableFactory` and are automatically generated upon registering a `TableFactoryDecorator` in the `Catalog`.

Furthermore, the `table_factory_` is automatically decoration at runtime based on the properties specified in the newly introduced `--table-properties` CLI argument.

All available properties can be listed with `--list-table-properties`.