[Util] Introduce expansion budget to A*.
Provide an expansion budget for the maximum number of expansions during
A* search.  This paves the way to extend A* to anytime search.

Introduce a new exception `m::ai::budget_exhausted_exception` to signal
the caller of `genericAStar` that the expansion budget was exhausted and
search was aborted.

Adapt `heuristic_search` to handle said exception.  Currently, there's a
TODO for that part.