[Scheduler] Add `Transaction` struct
The `Transaction` struct is added as part of the `Scheduler` struct. `Transaction`s will be used as to manage concurrency control, particularly in the context of setting and comparing multi-version timestamps within `Table`s.

The `DatabaseCommand`, `Reader`, and `QueryGraph` structs have been expanded to include a pointer to their corresponding `Transaction`.

A new `Transaction` object can be requested from the `Scheduler` with the `begin_transaction()` method. The resulting `unique_ptr` must from now on be passed by reference when calling `schedule_command()`. When the changes made by a `Transaction` are committed, the `commit()` method is used, which takes the `unique_ptr` of the `Transaction` as an argument and effectively closes the `Transaction`.

Another new method, `autocommit()`, is also added to the Scheduler. This method simplifies the execution of commands by automatically creating a new `Transaction`, scheduling the command, and committing the `Transaction` in a single method. This is the default behavior for executing commands until the transaction control language is implemented.

Finally, the `SerialScheduler::CommandQueue` now stores a reference to the `Transaction` of each command and orders queued commands based on the start time of their associated `Transaction`s. To support this functionality, the underlying container is changed from a `std::queue` to a `std::list`, allowing for insertions at any position within the queue. From the outside it now functions like a priority queue.