[IR] Pass HS config as template argument.
The `genericAStar` already takes a `SearchConfigConcept` template
parameter, that configures the search at compile time.

The `heuristic_search()` function, however, took a *partial* template
specialization of `genericAStar` as template parameter, with the
compile-time config baked into the search.  This makes accessing the
configuration unnecessarily difficult.

With this commit, we modify the template parameters of
`heuristic_search()` to take the search as a template class that is
*not* partially specialized anymore, and add a template parameter that
provides the compile-time configuration of the search.

This design is more idiomatic and makes using the config for
compile-time checks much easier.