[Catalog] Add pre- and post-optimizations
This commit introduces pre- and post-optimizations as new `Catalog` Components.

Pre-optimizations are implemented as callback functions that accept the QueryGraph as an argument, prior to its optimization.

Similarly, post-optimizations are defined as callback functions that receive the resulting `Producer` after the optimization process.

Notably, the commit also removes the `plan->minimize_schema()` operation at the end of optimization, registering it as a post-optimization step.