[ITest] Use multiprocessing to run tests in parallel
We refactor the script in order to enable running tests in parallel to
save execution time. For this, we use `multiprocessing.Pool` and create
a process inside this pool for each test case. We keep track of the
remaining jobs in the pool and update the progress bar accordingly.

Because the `tqdm` progress bar cannot be shared between processes, the
`report` methods had to be refactored. They now return a list of
messages to the caller instead of writing them on command line directly.
After a process is finished, the caller takes care of printing the
report messages.