[Util] Add `cheaper` and `decrease_key` counter
Add the `cheaper` and the `decrease_key` counter for the case that a cheaper path is found to a state that is only present in the regular queue but not in the beam queue.
In this case, the g-value of the state is decreased. Thus the `cheaper` counter has to be increased.
As the state is still present in a queue (the regular queue) and is not reopened, the `decrease_key` counter, reflecting the number of times that a g-value decreases of a states present in a queue, has to be increased.