[CMake] Fix static linking w/o V8.
When configuring a static library build with `-DWITH_V8=OFF`, there was
a CMake error that there is no rule to build `libv8_monolith`, that is
required by `libmutable_bundled`.  This was due to a mistake in the
configuration of the libraries included in `libmutable_bundled`.

This commit introduces a check for `WITH_V8` before adding the V8 and
Binaryen libraries to `libmutable_bundled`.