[Index] Add array-based index, use it by default
This commit adds `ArrayIndex` which is based on a sorted array.  For
now, this is the default index as it supports all data types.  `Sema`
now creates an index such that the key type matches the type of the
attribute that is indexed.  `CreateIndex` command also bulksloads the
index before adding it to the database.  Unit tests for `Sema` on
`CreateIndexStmt` were updated to use `ArrayIndex` as `IndexBase` is
`virtual` now and serves as super class to all future indexes.