[Scheduler] Use -1 as special Timestamp
Initially, we reserved 0 as a special timestamp for `Transaction`s and tuples.
For `Transaction`s a timestamp of 0 means the `Transaction` has not started executing statements yet.
For Tuples the value 0 is equivalent to infinity.

Since SQL does not support unsigned integers, this commit changes this special value from 0 to -1. This should help comprehension since -1 is often assigned a special meaning and we no longer skip over 0 when assigning timestamps.