[PhysOpt] Return found physical operator covering by value.
`PhysicalOptimizer::extract_plan()` now returns the found physical
operator covering by value instead of returning a reference to
it. This corresponds to the paradigm "the value is the boundary"
and prevents dangling references when the physical optimizer is
destroyed.
To enable this change, the physical plan tables no longer own the
found covering and the `Match<T>` objects no longer reference
their children, but both share the corresponding `MatchBase`s
by making use of `m::unsharable_shared_ptr`.