[ITest] More generous timeout per test case.
With the current timeout of 15 seconds, few long-running integration
tests timeout frequently.  This commit generously increases the timeout
to 60 seconds.  The goal of the timeout is not to ensure that
integration tests run fast, but to ensure they guarantee they terminate
eventually and the testing script (and hence CI) come to a result in a
reasonable amount of time.