[Util] Pools now do reference counting.
Massively refactor our pools `PODPool` and `Pool` to support reference
counting of pooled objects.

The pools are now *maps* from pooled object to the object's reference
count.  This allows for tracking whether we would leave any dangling
pointers around when deleting a pool.  Deleting a pool while the ref
count of a pooled object is greater than zero causes an abort in the
debug build.  This should give us good confidence that we don't leave
dangling references around.

This patch also prepares for garbage-collecting pooled objects that are
not being referenced anymore.  Making `Pooled` perform
garbage-collection is now a just a trivial change.  However, before we
can implement this, we must drop the implicit degeneration of a `Pooled`
to a raw pointer, thereby reducing the reference count while still
holding a reference through the raw pointer to the object.

This patch also prepares for *optional* `Pooled`s, that do not reference
any object.  I don't know yet how to design declaration and construction
in a convenient way...