[UTest] Fix #213: Increase delays in async test.
Increase delays in an asynchronous / multi-threaded test to reduce the
chance of odd scheduling.

The test randomly and infrequently failed before, as the scheduling
could be different than expected.  Our delays were simply not long
enough.  `fork()` can take a considerable amount of time...

This test **did not fail** because of data races or other
synchronization issues.