[Wasm] Add `IndexScan` operator
This operator covers a `Scan` followed by a `Filter`.  For this operator
to be applicable, an index must exist on the attribute referred to in
the filter condition and the filter condition must either be (1) an
equality predicate, (2) a one-sided range predicate, or (3) a two-sided
range predicate.  We also introduce two command line arguments.
`--scan-implementations` lets the user choose the scan implementations
to consider during physical optimization.  `--isam-batch-size` lets the
user choose the number of tuple ids that are communicated between host
and V8 during one callback.