[Index] Disable timer during index bulkload
Internally, bulkloading executes a query to retrieve all tuples of a
table.  If `--times` is set this internal query will also cause
execution times to be printed.  However, this causes problems with our
benchmarking script as too many measurements are extracted.  Disabling
the timer during the bulkload and re-enabling it afterwards solves this
issue as the execution times of the internal query are no longer
printed.