[Benchmark] Generate data of correct floating type
There was no distinction between generating `float` and `double` data.
We now correctly cast float data to `np.float32` which is compatible
with `float` in mutable and C++.  Further, it was not ensured that the
correct amount of distinct values where generated which is fixed by
using a hash set.