[Wasm] Fix #217: Handle duplicated hash table keys.
The previous fix in 6a3a870a was incorrect. Deduplication of
build and probe keys may yield an incorrect hash and comparison,
e.g., the condition `x = a AND x = b AND y = b` would be
incorrectly checked as `x = a AND y = b` due to deduplication.

The correct fix is not to `extract()` values from an environment,
but rather just `get()` them to not crash at duplicates.
Deduplication in the hash table schema, i.e., to prevent storing
the same value multiple times, has been already correclty
handled by hash tables by storing an indices mapping between
keys to access the hash table and its actually stored values
in the `key_indices_` member field.

This commit also includes respecitve integration tests for
duplicates in the build relation, in the probe relation, and
in both.