[Util] Fix deleter type of `unsharable_shared_ptr<T>`.
Prior to this commit, the custom deleter expects an argument of
type `const T*`. However, we allow conversion between unsharable
shared pointers of compatible types. The deleter type cannot be
changed in these cases which leads to invalid deleter accesses
(using the former deleter type) in `exclusive_shared_to_unique()`
after such a conversion. To circumvent the issue, we adapt the
custom deleter to receive an argument of type `const void*` and
cast the argument afterwards into the expected type.