[Wasm] Omit materialization in `wasm::SortMergeJoin` if possible.
Materialization can be omitted if the child operator
is a table scan and no sorting has to be performed.
In this case, we directly iterate over the underlying
table store instead of allocating a buffer.

Note that omitting materialization by reusing an
already filled buffer is possible in general, however,
our current implementation is restricted to passing
only tuples to parent operators.  Thus, this change
was easy to implement and will benefit our benchmarks
that assume an existing ordering on base relations.