[CI] Sanity check parallel unit test utility
Add a sanity check of the parallel unit test utility to CI.  This check
ensures that the utility correctly reprorts aborts, crashes, failures,
and successes.

Note that the result of this sanity run is added to the artifacts, but
not to the reports.  We do not want to clutter our GitLab test reports.