[Wasm] Fix computation of initial HT capacity.
Up to now, we potentially stored a value larger than a `uint32_t` in
`initial_capacity` leading to undefined behavior. We now properly cast
the value to a `uint64_t` and check if it fits into a `uint32_t`,
otherwise we return the maximal value of `uint32_t`.

Furthermore, we put the repetitive computation of the initial hash table
capacity into its own function.