[Wasm] Allocate hash table layout according to given schema.
The layout of each hash table entry now matches the given
schema instead of the given key indices. This enables
minimizing the storage requirement.

For example, if a hash table should be build on the composite
key with duplicates `(x,y,x)` with a value `z` as payload,
our hash table can be constructed in two ways:
1. Specify `[x,y,x,z]` as schema and `{0,1,2}` as key indices.
   Then, the value for `x` will be stored twice. Thus, this
   approach is not recommended, however, we do not enforce
   a deduplicated schema.
2. Specify `[x,y,z]` as schema and `{0,1,0}` as key indices.
   Then, the value for `x` will be stored only once, however,
   hash computation and key comparison will be performed as
   intended due to the specified key indices. Thus, this
   approach is recommended.