[IR] Insist that the PT is empty when filling it with the found plan
Currently we determine the initial upper bound for cost-based pruning by executing GOO on the PT. Filling the PT with the found plan, the PT is only updated when a cheaper partial plan is provided for the intermediate result.
Ensure that the PT is empty before reconstructing the found plan. This ensures that the PT contains the found plan after its reconstruction.