[Wasm] Add `like_prefix()` and `like_suffix()` optimizations.
The general `like()` computes a SQL LIKE using a dynamic
programming approach in O(n*m) runtime where `n` is the
string length and `m` is the pattern length. However, for
static patterns known at query compile time we can introduce
specialzations. We now support prefix and suffix expressions,
i.e. patterns of the form `.*%` and `%.*`, for which the
actual search at query runtime can be performed in O(n).