[Wasm] Fix `like_contains()` for 64-bit pointers.
Rewrite the inner loop's condition s.t. it is explictly
evaluated *lazily*, i.e., the pattern pointer is only
dereferenced if the pattern position is valid/non-negative.
Note that this rewrite is only possible since the initial
pattern position is always valid when trying to enter the
inner loop.

Honestly, this should have been an issue for 32-bit pointers
as well, but never occurred there.