[CMake] Append version number to third-party source directories
Whenever we update a third-party dependency, the project has to be
built from scratch due to potential conflicts.  This is particularly
painful in the CI, where caches have to be emptied.  Having
version-specific mostly source directories fixes this issue.  In some
cases the build directory of a third-party dependency may have to be
emptied.  However, we do not cache these in our CI.